requirements.txt
Django==5.0.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.2
python-decouple==3.8
Pillow>=11.0.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.10
django-filter==24.2
gunicorn>=21.2.0
